refactor: separate ProductEditForm business logic as hook
